Moved everything under BigEndianPatricia to make way for changes
Ignore-this: 3137bc8b1b1ddfb697d87fdd8492a1f9

darcs-hash:92850765d7fcab889a70378d7604382c98bd7053
