+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2008--2013, wren gayle romano.
+Copyright (c) 2008--2014, wren gayle romano.
 ALL RIGHTS RESERVED.
 
 Redistribution and use in source and binary forms, with or without
