+++ VERSION
@@ -1,5 +1,3 @@
-0.2.4 (2014.10.09):
-    - added Data.Trie.Internal.{match_,matches_}, Data.Trie.Base.{match,matches}
 0.2.3 (2010.02.12):
     - added Data.Trie.Internal.alterBy_
     - added Data.Trie.Internal.{contextualMap, contextualMap', contextualFilterMap, contextualMapBy}
