+++ bytestring-trie.cabal
@@ -1,5 +1,5 @@
 ----------------------------------------------------------------
--- wren gayle romano <wren@community.haskell.org>   ~ 2014.10.09
+-- wren gayle romano <wren@community.haskell.org>   ~ 2013.01.10
 ----------------------------------------------------------------
 
 -- By and large Cabal >=1.2 is fine; but >= 1.6 gives tested-with:
@@ -8,12 +8,12 @@ Cabal-Version:  >= 1.6
 Build-Type:     Simple
 
 Name:           bytestring-trie
-Version:        0.2.4
+Version:        0.2.3.2
 Stability:      provisional
 Homepage:       http://code.haskell.org/~wren/
 Author:         wren gayle romano
 Maintainer:     wren@community.haskell.org
-Copyright:      Copyright (c) 2008--2014 wren gayle romano
+Copyright:      Copyright (c) 2008--2013 wren gayle romano
 License:        BSD3
 License-File:   LICENSE
 
@@ -74,12 +74,12 @@ Flag bytestringInBase
 ----------------------------------------------------------------
 Library
     Hs-Source-Dirs:  src
-    Exposed-Modules: Data.Trie
-                   , Data.Trie.Internal
-                   , Data.Trie.Convenience
-    Other-Modules:   Data.Trie.BitTwiddle
-                   , Data.Trie.ByteStringInternal
-                   , Data.Trie.Errors
+    Exposed-Modules: Data.Trie.BigEndianPatricia.Base
+                   , Data.Trie.BigEndianPatricia.Internal
+                   , Data.Trie.BigEndianPatricia.Convenience
+    Other-Modules:   Data.Trie.BigEndianPatricia.BitTwiddle
+                   , Data.Trie.BigEndianPatricia.ByteStringInternal
+                   , Data.Trie.BigEndianPatricia.Errors
     Build-Depends: binary
     
     -- I think this is all that needs doing to get rid of the warnings?
